% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRASSBS.R
\name{calc_basin_stats}
\alias{calc_basin_stats}
\title{Wrapper. Function reads in a table of coordinates with a unique ID 'UID' field and
calculates raster statistics for a input GRASS raster. Statistics are calculated
using GRASS r.univar and written to the ouput directory as txt files.}
\usage{
calc_basin_stats(basin_df, stat_rast, procs, out_dir)
}
\arguments{
\item{basin_df}{A data frame with x-y coordinates --labeled as x & y-- of pour points to delineate and a unique identifier feild --labeld UID--.}

\item{stat_rast}{Name of input GRASS raster to calculate basin statistics as string, e.g., 'slope'. Assumed present in GRASS initialized environment.}

\item{procs}{Number of threads to run in parallel, recommend no more than 90\% of system total.}

\item{out_dir}{Output directory for GARSS r.univar output txt files with basin statistics.}
}
\value{
Nothing, however output directory is populated with basin statistics as txt files.
}
\description{
Wrapper. Function reads in a table of coordinates with a unique ID 'UID' field and
calculates raster statistics for a input GRASS raster. Statistics are calculated
using GRASS r.univar and written to the ouput directory as txt files.
}
